+++ rollup.config.js
@@ -35,5 +35,15 @@ export default {
 			open: true
 		})
 	],
-	external: ['classnames', 'tslib', 'lodash.clonedeep', 'lodash.clone', 'react-select', 'react-toastify', 'axios']
+	external: [
+		'classnames',
+		'tslib',
+		'lodash.clonedeep',
+		'lodash.clone',
+		'react-select',
+		'react-toastify',
+		'axios',
+		'react-select',
+		'react-select/creatable/dist/react-select.esm.js'
+	]
 };