+++ src/ui/inputTextarea/InputTextarea.tsx
@@ -8,13 +8,14 @@ import { IRsFormControl } from '../form/FormControl';
 
 import { Box } from '../box/Box';
 import clone from 'lodash.clone';
+import { renderErrors } from '../../utils/internal';
 
 export interface InputTextareaProps
 	extends ICommon.PaletteProps,
 		Omit<ICommon.HtmlElementProps, 'display'>,
-		Omit<TextareaHTMLAttributes<HTMLTextAreaElement>, 'onChange'> {
-	//TextInput Props
-	boxRef?: React.RefObject<HTMLDivElement>;
+		Omit<TextareaHTMLAttributes<HTMLTextAreaElement>, 'onChange'>,
+		ICommon.MarginProps {
+	elementRef?: React.RefObject<HTMLDivElement>;
 	look?: 'standard' | 'filled' | 'outlined' | string;
 	noAutocomplete?: boolean;
 	autocompleteType?: ICommon.AutoCompleteType | string; // Defaults to "on"
@@ -32,14 +33,14 @@ export interface InputTextareaProps
 
 	onFocus?: (event: React.FocusEvent<HTMLTextAreaElement>) => void;
 	onBlur?: (event: React.FocusEvent<HTMLTextAreaElement>) => void;
-	onChange?: (event: React.ChangeEvent<HTMLTextAreaElement>, control?: RsFormControl<IRsFormControl>) => void;
+	onChange?: (value: string, event: React.ChangeEvent<HTMLTextAreaElement>) => void;
 }
 
 const InputTextarea: React.FC<InputTextareaProps> = (props) => {
 	const {
 		id,
 		look,
-		boxRef,
+		elementRef,
 		control,
 		updateControl,
 		noAutocomplete,
@@ -51,8 +52,40 @@ const InputTextarea: React.FC<InputTextareaProps> = (props) => {
 		minLength,
 		maxLength,
 		useFloatingPlaceholder,
+		m,
+		mt,
+		mr,
+		mb,
+		ml,
+		mx,
+		my,
+		margin,
+		marginTop,
+		marginRight,
+		marginBottom,
+		marginLeft,
+		marginX,
+		marginY,
 		...textareaProps
 	} = props;
+
+	const boxMarginProps = {
+		...(m && { m }),
+		...(mt && { mt }),
+		...(mr && { mr }),
+		...(mb && { mb }),
+		...(ml && { ml }),
+		...(mx && { mx }),
+		...(my && { my }),
+		...(margin && { margin }),
+		...(marginTop && { marginTop }),
+		...(marginRight && { marginRight }),
+		...(marginBottom && { marginBottom }),
+		...(marginLeft && { marginLeft }),
+		...(marginX && { marginX }),
+		...(marginY && { marginY })
+	};
+
 	const textareaRef = useRef<HTMLTextAreaElement | null>(null);
 
 	const [formControl, setFormControl] = useState<RsFormControl<string> | undefined>(control);
@@ -62,25 +95,28 @@ const InputTextarea: React.FC<InputTextareaProps> = (props) => {
 	}, [control]);
 
 	async function changeHandler(event: React.ChangeEvent<HTMLTextAreaElement>) {
+		// Required to persist in React 16.X but not 17.X. Otherwise the await() will lose the object
 		event.persist();
+
+		if (onChange) onChange(event.target.value, event);
+		if (!control) return;
+
 		const target = event.target;
 
 		const startPosition = target.selectionStart || 0;
 		const endPosition = target.selectionEnd || 0;
 		const updated = clone(control);
-		if (updated) {
-			updated.value = target.value;
-			if (updated.value.length === 0) {
-				updated.clearErrors();
-			} else {
-				await updated.validate();
-			}
-			setFormControl(updated);
-			if (updateControl) updateControl(updated);
+
+		updated.value = target.value;
+		if (updated.value.length === 0) {
+			updated.clearErrors();
+		} else {
+			await updated.validate();
 		}
+		setFormControl(updated);
+		if (updateControl) updateControl(updated);
 
 		target.setSelectionRange(startPosition, endPosition);
-		if (onChange) onChange(event, updated);
 	}
 
 	function getAutocompleteType(): string {
@@ -113,21 +149,6 @@ const InputTextarea: React.FC<InputTextareaProps> = (props) => {
 		);
 	}
 
-	function renderErrors() {
-		if (!control) return;
-		const errorNodes: React.ReactNode[] = [];
-		const errors = control.errors;
-		for (let index = 0; index < errors.length; index++) {
-			const errorMessage = control.getErrorMessage(errors[index]);
-			errorNodes.push(
-				<div key={`${index}Error`} className={'rsInputTextareaErrorMessage'}>
-					{errorMessage}
-				</div>
-			);
-		}
-		return errorNodes;
-	}
-
 	function focusInput() {
 		if (textareaRef && textareaRef.current) textareaRef.current.focus();
 	}
@@ -136,14 +157,15 @@ const InputTextarea: React.FC<InputTextareaProps> = (props) => {
 		<Box
 			id={id}
 			className={classNames('rsInputTextarea', className, look)}
-			elementRef={boxRef}
+			elementRef={elementRef}
 			color={props.color}
 			bgColor={props.bgColor}
+			{...boxMarginProps}
 		>
 			<Box className={'inputTextareaContainer'} onClick={focusInput}>
 				{renderTextarea()}
 			</Box>
-			{renderErrors()}
+			{renderErrors(props.control)}
 		</Box>
 	);
 };